requirements.txt
tqdm==4.64.1
jupyter==1.0.0
typer==0.7.0
pandas==1.3.5
numpy==1.24.2
spacy==3.4.3
--extra-index-url https://download.pytorch.org/whl/cu113
torch==1.12.1+cu113
transformers==4.25.1
datasets==2.7.1
evaluate==0.3.0
nltk==3.8.1
scikit-learn==1.2.2
gensim==4.3.1
neptune-client==0.16.13